+++ packages/core/src/inIdleCallback.ts
@@ -28,6 +28,24 @@ class InIdleCallbackProducer implements Producer<IdleDeadline> {
 /**
  * Takes optional IdleRequestOptions, and returns a lazy promise that resolves
  * with `IdleDeadline` in an idle callback.
+ *
+ * To defer execution of a callback, use
+ *
+ * ```
+ * inIdleCallback().pipe(map(() => ...))
+ * ```
+ *
+ * To make an existing lazy promise settle in an idle callback, pipe it though
+ *
+ * ```
+ * finalize(inIdleCallback)
+ * ```
+ *
+ * To limit this to only when the promise resolves, use
+ *
+ * ```
+ * map((value) => inIdleCallback().pipe(map(() => value)))
+ * ```
  */
 export const inIdleCallback = (
   options?: IdleRequestOptions,